Move prompt/checks to models/index.ts
